+++ Controller/DuplicateModController.php
@@ -10,6 +10,40 @@ namespace Kanboard\Plugin\DuplicateMod\Controller;
 class DuplicateModController extends \Kanboard\Controller\PluginController
 {
     /**
+     * Show one of the settings pages of the DuplicateMod plugin.
+     *
+     * @return HTML response
+     */
+    public function showConfig()
+    {
+        // !!!!!
+        // When I want to add new config options, I also have to add them
+        // in the WeekHelperHelper.php in the getConfig() Method !
+        // !!!!!
+        $this->response->html($this->helper->layout->config('DuplicateMod:config/duplicatemod_config', $this->helper->duplicateModHelper->getConfig()));
+    }
+
+    /**
+     * Save the setting for DuplicateMod.
+     */
+    public function saveConfig()
+    {
+        $form = $this->request->getValues();
+
+        $values = $this->helper->duplicateModHelper->generateConfig($form);
+
+        $this->languageModel->loadCurrentLanguage();
+
+        if ($this->configModel->save($values)) {
+            $this->flash->success(t('Settings saved successfully.'));
+        } else {
+            $this->flash->failure(t('Unable to save your settings.'));
+        }
+
+        return $this->response->redirect($this->helper->url->to('DuplicateModController', 'showConfig', ['plugin' => 'DuplicateMod']), true);
+    }
+
+    /**
      * Duplicate a task without asking first
      *
      * @access public