+++ Locale/de_DE_du/translations.php
@@ -2,4 +2,6 @@
 
 return array(
     'Duplicate (instant)' => 'Duplizieren (sofort)',
+    'DuplicateMod configuration' => 'DuplicateMod Einstellungen',
+    'Prefix for duplicated task (leave empty for none).' => 'Prefix für den duplizierten Task (leer lassen für keinen).',
 );