+++ Template/config/duplicatemod_config_sidebar.php
@@ -1,0 +1,3 @@
+<li <?= $this->app->checkMenuSelection('DuplicateModController', 'showConfig') ?>>
+    <a href="/duplicatemod/config"><?= t('DuplicateMod configuration') ?></a>
+</li>